requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.34
PyJWT==2.3.0
psycopg2-binary==2.9.9
werkzeug==3.0.4
